# docs/api/storefront/openapi.json
{
...
  "info": {
    "description": "OpenAPI contract for the public Matjero Storefront API.",
    "title": "Matjero Storefront API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/storefront/categories": {
      "get": {
        "description": "Returns the public category tree for the resolved store. Each node carries its parent slug so a client can rebuild the hierarchy.",
        "operationId": "listStorefrontCategories",
        "parameters": [
          {
            "description": "Override locale resolution",
            "in": "query",
            "name": "locale",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "items": {
                      "items": {
                        "properties": {
                          "description": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "parent_slug": {
                            "type": "string"
                          },
                          "product_count": {
                            "format": "int64",
                            "type": "integer"
                          },
                          "slug": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "name",
                          "product_count",
                          "slug"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "items"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Public category collection"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Invalid query parameters"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Storefront or resource not available"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "List public categories",
        "tags": [
          "Storefront",
          "Categories"
        ]
      }
    },
    "/v1/storefront/categories/{slug}": {
      "get": {
        "description": "Resolves a category slug within the store resolved from the request host. A slug belonging to another store returns 404.",
        "operationId": "getStorefrontCategory",
        "parameters": [
          {
            "description": "Override locale resolution",
            "in": "query",
            "name": "locale",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Public category slug",
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "category": {
                      "properties": {
                        "description": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "parent_slug": {
                          "type": "string"
                        },
                        "product_count": {
                          "format": "int64",
                          "type": "integer"
                        },
                        "slug": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "name",
                        "product_count",
                        "slug"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "category"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Public category"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Invalid query parameters"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Storefront or resource not available"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Get a public category by slug",
        "tags": [
          "Storefront",
          "Categories"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}